Removed core::ManualGraph since allowing the user to specify the type of Graph::Vertex is redundant. Instead, the user can use the weight if needed, making it a tuple if more than one weight is needed.

+++ tests/mock_graph/mod.rs
@@ -17,7 +17,7 @@ use graphene::core::Directedness;
 #[derive(Copy, Clone, Eq, PartialEq, Debug, Hash)]
 pub struct MockVertex
 {
-	pub value: u32
+	pub value: usize
 }
 
 #[derive(Debug, Clone, PartialEq)]
